+++ packages/stream_video_push_notification/lib/src/stream_video_push_notification.dart
@@ -32,6 +32,11 @@ class StreamVideoPushNotificationManager implements PushNotificationManager {
   static create({
     required StreamVideoPushProvider iosPushProvider,
     required StreamVideoPushProvider androidPushProvider,
+    @Deprecated(
+      "Caller customization is deprecated as it was not fully compatible with iOS "
+      "(foreground calls only). Use 'display_name' custom field in the call instead. "
+      "See details: https://getstream.io/video/docs/flutter/advanced/incoming-calls/customization/#display-name-customization",
+    )
     CallerCustomizationFunction? callerCustomizationCallback,
     @Deprecated(
         'Background handler is no longer needed for terminated state ringing on iOS.')
@@ -220,6 +225,11 @@ class StreamVideoPushNotificationManager implements PushNotificationManager {
   final StreamVideoPushProvider iosPushProvider;
   final StreamVideoPushProvider androidPushProvider;
   final StreamVideoPushParams pushParams;
+  @Deprecated(
+    "Caller customization is deprecated as it was not fully compatible with iOS "
+    "(foreground calls only). Use 'display_name' custom field in the call instead. "
+    "See details: https://getstream.io/video/docs/flutter/advanced/incoming-calls/customization/#display-name-customization",
+  )
   final CallerCustomizationFunction? callerCustomizationCallback;
   final bool registerApnDeviceToken;
   late SharedPreferences _sharedPreferences;
@@ -315,6 +325,7 @@ class StreamVideoPushNotificationManager implements PushNotificationManager {
     String? nameCaller,
     bool hasVideo = true,
   }) {
+    // ignore: deprecated_member_use_from_same_package
     final customData = callerCustomizationCallback?.call(
       callCid: callCid,
       callerName: nameCaller,
@@ -342,6 +353,7 @@ class StreamVideoPushNotificationManager implements PushNotificationManager {
     String? nameCaller,
     bool hasVideo = true,
   }) {
+    // ignore: deprecated_member_use_from_same_package
     final customData = callerCustomizationCallback?.call(
       callCid: callCid,
       callerName: nameCaller,